[netlink] Only send Ack or Done message if needed
Previously, we responded to all GET requests with a Done message and
all non-GET requests with an Ack message. This change updates the
library to only send Done if the GET request was a dump (had the
NLM_F_DUMP flag set), and only send an Ack if the request requested
one (had the NLM_F_ACK flag set).

As a result of this change, some starnix gvisor rtnetlink tests had
to be disabled as they blocked forever. They blocked forever because
the tests would follow the pattern of:
  1) Send request without `NLM_F_ACK` set (don't send Ack).
  2) Read response from message.

Before this change, the implementation would unconditionally respond to
the tests' requests with an Ack and the test would read that message and
fail as it was not what the test expected. With this change, the tests
that do not request an Ack get no response as the requested operations
are not yet supported. This results in the read calls to the (blocking)
rtnetlink socket to never return.

Bug: b/283827094
Change-Id: If85dc0b0eb20a81d78a45f32f1cb49e6365ff29e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867958
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>