[profiler] Added process watcher
This cl implements a process watcher for the cpu profiler. It enables it
to monitor a thread and be notified when a monitored process starts a
new thread so that we can monitor it as well.

Since we're now passing around an async dispatcher anyways, this cl
also moves away from sampling on a std::thread and instead samples using
the async loop. This simplifies some locking, since we'll now never access
the data members concurrently and I did not observe a performance
difference in sample overhead.

Change-Id: I66b1ba0ad6ab76823b79c7b4dc357e84d4bd1327
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862145
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Brian Hamrick <bhamrick@google.com>