[wlan] Introduce the `event` module and the `Handler` trait.
This change introduces the `event` module and its core `Handler` trait
in hardware simulator.

This is the first in a series of changes that implement a new event
handler framework for writing hardware simulator tests. The `Handler`
trait, its combinators, and its adapters are the primary mechanism that
powers this API. These additions are based on a series of changes that
include a complete port of tests. Some documentation refers to these
subsequent changes.

Test: Unit tests and hardware simulator tests.
Change-Id: I1f009a03403df5696362decf7211ba777083d7fa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/811381
Reviewed-by: Dylan Swiggett <swiggett@google.com>
Commit-Queue: Sean Olson <seanolson@google.com>