[pkg] Remove meta.far.merkle
The support for generating and reading a meta.far.merkle file is
being removed from pm and package-tool. This file used to be used
for collecting information about packages during publish/build,
but we now use the package_manifest.json for this purpose.

Bug: 101304, 128626
Change-Id: I4c3e1f2b5b9a9518843afdfbdef83db8b5d9e39e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867299
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Craig Stout <cstout@google.com>
Reviewed-by: Andrew Pollack <andrewpollack@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>