[lockdep] Add a missing annotation
Add a missing annotation to the `lock()` method of lockdep's
`Lock<...>` wrapper.  When using instrumented locks with lockdep,
users end up wrapping their lock type inside of a lockdep wrapper,
which ends up encapsulating the instance of the lock, using the type
specified by the user.  For example:

```
struct MyStruct {
  LOCK_DEP_INSTRUMENT(MyStruct, SpinLock) my_spinlock_;
  uint32_t my_int_ TA_GUARDED(my_spinlock_);
}
```

Under the hood, the type is _not_ `SpinLock`.  Instead, it is
`lockdep::LockDep<...>, which derives from ``lockdep::Lock<...>`,
passing the SpinLock type down the hierarchy where it is instantiated
by `Lock<...>`.  The `Lock<...>` class is annotated to be a
capability, which is why we can declare `my_int` to be guarded by it.

Generally speaking, people typically use lockdep's RAII `Guard` (as
they should) when using their locks.  The `Guard` instance is
properly annotated to indicate that it is acquiring and releasing the
capability represented by the `Lock<...>` instance.

`Lock<...>` also provides access to the underlying encapsulated lock
via a `lock()` method, which returns a reference to the encapsulated
`LockType` instance.  This feature lets users access their specific
underlying lock type, allowing them to access lock features which
lockdep does not intrinsically understand.  These features could be
things not really related to memory protection, like a lock name used
for debugging, but could also include locking routines such as
`TryBeginReadTransactionDeadline`.

Whenever users manipulate their locks via the `lock()` accessor, they
are fundamentally manipulating the same capability that they would
have been manipulating via a `Guard`.  Right now, however, there are
no annotations on `lock()` which lets the compiler know this.  The
compiler just sees someone fetching an arbitrary lock using a method
on an object which happens to be (what appears to be) a different
capability.  So, the following code would not work, even though it
probably should.

```
void Foo(MyStruct& s) {
  s.my_spinlock_.Acquire();
  s.my_int_ += 5;
  s.my_spinlock_.Release();
```

Adding proper `__TA_RETURNS_CAPABILITY` annotations to `lock()`,
indicating that the reference it returns represents the same
capability as the instance itself fixes this issue.

Change-Id: I31f4497895f2bac170a494bf03d2d4c34e960b58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/866690
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>