[amlogic-display] Store image pointers instead of handles / IDs.
This change is similar to fxrev.dev/867723 but it's for amlogic-display
driver. `capture_active_id_` is stored as handle types (uint64_t)
which hides the fact that it is actually a raw pointer to ImageInfo.

Instead of storing a handle, we should store a raw pointer to
ImageInfo and convert them to / from uint64_t handles only at the
banjo API boundary.

Test: No-op. amlogic-display-test
Bug: 128704
Bug: 128653
Change-Id: I6947dfb0a89fdf14d4218d0ef90af5011143574c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868378
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yilong Li <liyl@google.com>