[bazel_sdk] Fix ArgumentScope on python 3.11
With python 3.11, the Enum class implementation was changed in a way
that breaks the usage in ArgumentScope. The `EXPLICIT` value is
converted from an empty  tuple to an object in a way that causes the `in`
operator on it to fail. We need to mixin `tuple` to the ArgumentScope
type to ensure that the enum values can act as tuples.

Change-Id: I0cd7d65e79c28a2d403ba3a7c47f762b864d9ca1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/866524
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>