[concurrent] Make the SeqLock APIs easier to hold correctly.
Change the SeqLock APIs exposed by libconcurrent to be easier to use
correctly. In particular, make the memory-ordering synchronization
approach (either acq/rel or fence) a part of the lock type itself.
Then, introduce a new payload wrapper type (`SeqLockPayload<T,
LockType>`) which is an thin layer on top of a
`WellDefinedCopyable<T>` wrapper.  The new wrapper type demands that
the user pass the type of the lock the wrapper is meant to work with,
and exposes `Read/Update` methods which do not give the user any
control over the type of synchronization implemented.  Instead, they
simply "do the right thing" based on properties of the lock type.

Additionally:
1) Update the lockdep wrappers present in `kconcurrent` in order to
   work with the new type-linked approach.
2) Update the existing users of SeqLock in the kernel.
3) Update the documentation in `libconcurrent`'s readme to demonstrate
   the new (harder to mess up) way to hold a SeqLock in the kernel.

Change-Id: Icce22b89de6335f8006323d884605116b0a30d22
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/810594
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>