[kernel][jtrace] Add protection against nullptr records.
When JTRACE's "last-per-cpu-event" feature is enabled, it introduces a
rare edge case which can cause a double panic.  If we encounter a
panic, and CPU X has _never_ recorded any events, the record for that
CPU is full of nothing but zeros.

If we attempt to print this record, it means that pointers to what
should be valid string literals (such as the tag), or the
file-function-line record, will be nullptr.  Attempts to render this
information will trigger a panic, usually when we are already
attempting to dump the trace data during a panic.

Add protection against this behavior by always checking for nullptr in
the SafeString helper class, as well a explicitly checking the FFL
record during a dump.

Change-Id: Id43433b421897f3cc10f7f6480bf0487895ab655
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868882
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>