[botanist] Copy ffx logs to outputs if copying all outputs fails.
Sometimes the ffx daemon fails to shutdown properly leaving a
daemon.sock in the outputs dir which causes an error when trying to copy
the dir to the CAS outputs because it's an unknown filetype. For
debugging purposes, try to copy the ffx logs dir anyway even if we get a
failure trying to copy over the entire directory.

Bug: 128608
Change-Id: Ie5eba440d799204f826b08a160ffe522c93954d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869276
Reviewed-by: Oliver Newman <olivernewman@google.com>
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>