[netlink] Indicate group when sending multicasts
If requested by the application, the kernel is expected to set the
source address of the message. All messages initiated by the netlink
subsystem have PID 0 to match Linux. The group field is set to the
multicast group the message was directed to (if the message was a
multicast).

Bug: b/283827094
Cq-Include-Trybots: luci.fuchsia.try:fuchsia-coverage
Change-Id: Ie5ccc6e102c8c2eb136199209008e25645798ac1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868681
Fuchsia-Auto-Submit: Ghanan Gowripalan <ghanan@google.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Antonio Linhart <antoniolinhart@google.com>