[dhcp/client] Implement Renewing state
Implements unicasting renewal requests to the DHCP server we originally
got the lease from and handling the response. If we never hear back from
the server, then we transition to Rebinding, but the Rebinding state
itself is still not finished, so instead we transition to Init as if
the lease had expired.

This means that we have worse behavior when there are multiple potential
DHCP servers on the network, but still allows us to renew "correctly"
in most cases without ever needing to hang onto a lease beyond its
expiration time.

Bug: 81593
Fixed: 124593
Fixed: 125443
Change-Id: If8f3cd3433966c3caff4d56cfed4d1b9952f62f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/851800
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Alex Konradi <akonradi@google.com>
Commit-Queue: Evan Wildenhain <wildenhain@google.com>