[kernel][riscv] unmask specific external interrupts earlier per cpu
Unmask software (IPI) and external interrupts earlier in the boot
process per cpu to make sure by the time they're needed the cpu is ready
to handle them. This is safe early on because the sstatus.sie bit acts
as the main overriding mask, which isn't set until the cpu is ready to
schedule.

This fixes a wedged kernel boot that happens occasionally where the boot
cpu gets into a place where it is blocked waiting for timers or IPIs to
kick the scheduler, but it never occurs.

Change-Id: Id014ca3b69b52c0899b4d5ba34c190e208f095e6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868886
Fuchsia-Auto-Submit: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>