[bt][obex] Define type for the `Type` header
The Type Header is used to describe the type of Obex object. It is
based on a fixed set of media types (MIME) and is represented as a
null-terminated ascii string in Obex. This differs slightly from the
canonical OBEX string which is represented as a utf-16 null-terminated
string.

Define the MimeType for representing this header and some conversions
to/from a typical Rust string.

Bug: 120012
Test: fx test bt-obex-tests
Change-Id: I21cf58c6617bec077f1fea3d54d7cbe456c6c48a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867458
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>