[dart][inspect] Remove test
As part of the work to move dart OOT I think we can remove this test.
This test depends on an in-tree Rust component. Rather than rewriting
that component OOT we can remove the test.

The library that the test is testing is only used by the inspect dart
integration test. That integration test arguably serves as a test for
the library itself given its current limited functionality and use.

Bug: b/240154053

Change-Id: I754f81503640537ea97449a964e32dc3880a1a25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870443
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>