[fuchsia-controller] Breaking change: API rename.
This takes the "open_*" functions and replaces them with "connect_*"
functions to clearer demonstrate what is happening with the FIDL
connections. This also renames "open_notification_fd" to
"connect_notification_fd" since this can also be done more than once
without explicitly opening a file descriptor every time.

The impetus for this change is from RFC comments about
fuchsia-controller, so is likely a prerequisite for landing this in the
SDK.

Change-Id: Ic87baf59279096a48d0d975f1edf5b1475a6e7d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870039
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>