[superproject] Roll fuchsia [netstack] Increase send-ready timeout in test
Increase the timeout on waiting for send-ready after calling connect
used in the TestBridge unit test from 1s to 3s. The test expects
send-ready to be asserted when the bridge is up within the timeout,
and relies on the timeout occurring after the bridge is torn down to
determine that the connect fails. When the test is run in infra,
sometimes it takes more than 1s for the connect to succeed and
send-ready to be true, which causes the test to spuriously fail.

Multiply: netstack-bridge-gotests
Original-Fixed: 128191
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869676
Original-Revision: 1218d08db4252588a6d2708bd5fc979c1f1a9687
GitOrigin-RevId: 91cce14e80a7c08f9cb73e8a00735bfff04797cc
Roller-URL: https://ci.chromium.org/b/8778376176396359121
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ic83be777f1959173b1aaca5c3e905e392d9e82e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870696