[netlink] Support Dump RTM_GETROUTE e2e
This change allows netlink sockets/clients to perform the `RTM_GETROUTE`
request with `NLM_F_DUMP` flag set. The behavior follows `ip route
list`, where if the address_family is set to AF_UNSPEC, the v4 routes
will be returned prior to the v6 routes.

The Starnix/gvisor test that makes use of GETROUTE, `GetRouteDump`,
continues to fail as it did previously, but is now closer to passing.
After a change to the netlink-packet library, another CL will be created
to make this test pass, which will also require a change of the `table`
field in the RouteMessage header.

Fix: b/283137647

Change-Id: I83e56e525d881c7c1c9d14766fca5829f0d10db7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868396
Commit-Queue: Antonio Linhart <antoniolinhart@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>