[fxblob] Write blob benchmarks
WriteBlobWithFidl measures the time is takes blobfs to write a single
blob of a particular size via fuchsia.io. WriteBlobWithBlobWriter
measures the time it takes fxblob to write a blob of a particular size
via the new write API. We test WriteBlobWithFidl and
WriteBlobWithBlobWriter against small and large blob sizes to measure
the effects of back pressure.

WriteRealisticBlobsWith(Fidl/BlobWriter) measures the time it takes
blobfs/fxblob to write a set of varying-sized blobs. The client can
concurrently write up to two blobs at a time in this benchmark.

Bug: 125334
Change-Id: Idc56b23d78226ba5abe1de58f01eea293618a4e8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867578
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Commit-Queue: Nikita Jindal <nikitajindal@google.com>