[netstack3] Don't apply window scaling in SYN-RCVD
We don't advertise a window size that is scaled, however, we mistakenly
used a scaled window size to calculate the acceptable window for any
incoming segment. This made a lot of segments that should be outside the
window considered by the state machine as inside the window.

Change-Id: Ic2cf51acd0b12fd8df88171ed631718a57285262
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/870838
Reviewed-by: Nick Brown <nickbrow@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>