[fan-controller] Fix flake
For the second fan device added of the same client type, Watch() has
already been called and will not be called again. Thus, waiting for
the Watch() call cannot be an indicator that the fan device has been
registered. Use a separate condition for this.

Bug: 128656

Multiply: fan-controller-test

Change-Id: I4aa21ae0fc34446f205433afe364fc1a773aa149
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868938
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Ruby Zhuang <rdzhuang@google.com>