[platform_config][diagnostics] Write logs to serial from Archivist not Console
This change will cause the Archivist to log to serial instead of
console.

The new configuration is based on monikers. The configuration reflects
this based on analysis of what tags map to which components.

Based on previous discussion, we are also extending the set of
components that write to serial by including everything under the
/bootstrap realm.

NO_IFTTT=we don't need these changes in archivist default config in gni which are inherited by archivists for testing and the embedded archivist

Bug: 100486
Change-Id: I68dd19adac57545a8a32e7932774c1b1cea2ce42
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/867197
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>