[network/tests] Tun socket benchmarks with tracing
Add a variant of the tun socket benchmarks that collects a trace during
the benchmarks for further analysis.

Add trace events to the benchmarks to show the duration of each
test group, test case, and read/write operation.

The dart wrappers of benchmarks for the purpose of generating a trace
have been collected into a new file to enable reuse of the existing
function for running the benchmark and downloading the trace.

Bug: 125658
Change-Id: Id15826ca7a621c37307e8318de8da02c46679482
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/862228
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Mark Seaborn <mseaborn@google.com>