[gpt] Provide hardware.block.volume.VolumeManager
Implements a subset of the VolumeManager interface for the GPT driver,
allowing runtime allocation of GPT partitions.

The GPT device tree now contains a `gpt` node which offers the
VolumeManager protocol.  Note that partitions are *not* moved to
children beneath the `gpt` node, since doing so would require extensive
refactoring of the system (i.e. removing hard-coded GPT paths).  The
`gpt` device is a peer of the partition devices.

Newly allocated partitions are immediately bound, and other partitions
are unaffected (which is unlike the `gpt` host tool which rebinds all
GPT drivers upon any header modifications).

This is intended to support the storage-benchmark tests on systems which
do not have an FVM.

This change required refactoring the GptDevice library to receive a
BlockDevice rather than a BlockProxy, since we can't create a BlockProxy
for driver->driver communication.  The GPT driver provides an
implementation of the BlockDevice interface which directly communicates
using the inter-driver Banjo block protocol.

Bug: 126938
Change-Id: Ibe6717cf458dfad7fcc35f10fe8dc33adde3b103
Tested: Manually paved a NUC to ensure paving works with this change.
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868940
Size-Review: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>