[debugger] Catch failures stepping through PLT.
When there is a problem with the breakpoint that is step throught a PLT
call, currently we'll just resume execution.

This patch adds a breakpoint at the return address of the PLT call to
catch the various errors that can occur handling the PLT trampoline.
While this ends up stepping over the PLT call rather than into it, at
least execution is not implicitly resumed. In most places we step over
PLT trampolines and something fails (see bug 128935) the user is
actually trying to step over anyway so the problem is not visible.

Bug: 127391

Change-Id: I4d5a71b46b88d0aaa38d50ff9408c8d7f612fb4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/869899
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>