[starnix] Send SIGALARM in setitimer() and implement alarm()
This updates out itimer implementation to generate SIGALARM signals when
firing and also to properly compute the remaining time value when the
timer is queried. This is implemented by running asynchronous tasks on
the kernel thread pool's async executor and using the async executor's
timer heap to track the different timeouts.

Only ITIMER_REAL is implemented as we do not have a mechanism for
tracking the clocks needed by ITIMER_VIRTUAL or ITIMER_PROF. Querying
either of these timers succeeds and returns the disarmed state. We also
support 'setting' the timer to the disarmed state as this is used by
gVisor tests during setup to suppress SIGPROF signals.

This also supports the legacy x86_64 alarm() syscall which is similar to
setitimer(ITIMER_REAL, ...) except that it has slightly different output
behavior.

Change-Id: I005bd1f96b655611ad1a3456d7a448c040943ff4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/868194
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Adam Barth <abarth@google.com>