[netstack3] Fix a TCP data corruption bug
The readable region of the send buffer is only a portion of the entire
buffer, previously we bump snd.nxt without considering short reads are
possible on send buffers, this leads to wrong sequence numbers. It is
uncovered by a performance degradation due to retransmissions because
this bug caused gaps in the sequence numbers which are supposed to be
next to each other, it turns out to be a correctness issue.

Fixed: 128850
Change-Id: Ibeb97a29fe0a857eb7b504cd0c45afd2a62b4ca5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872016
Reviewed-by: Alex Konradi <akonradi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>