[pkg] Delete fuchsia.pkg/PackageCache.Open
Open depends on the dynamic index which is being removed.

Clients are migrated to Get which will succeed as long as the blobs
are cached, regardless of the state of the dynamic index.

As a result, the cache fallback for resolves will now succeed as long
as the blobs are cached, which could be observed in the case where
ephemeral resolution of a cache package succeeds and then, without the
device rebooting or GC being triggered, a later ephemeral resolution
fails. This behavior change was approved in RFC-0217.

Fixed: 124713
Change-Id: I32597472ee62aae7bc2ecac243e78137f0792f2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/827996
Commit-Queue: Ben Keller <galbanum@google.com>
API-Review: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Reviewed-by: Ali Zhang <alizhang@google.com>
Fuchsia-Auto-Submit: Ben Keller <galbanum@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>