[gvisor_syscall_tests] Skip ClosedWriteBlockingSocket for ns3
This test has been flaky for a while despite workarounds. The fact the
test is dependent on the concrete implementation of the socket buffers
makes it hard to device a method for all platforms (linux, gvisor, ns2
and ns3). A previous attempt that accommodates fuchsia netstacks caused
flakes in gVisor tests. Note that this test is also skipped for ns2,
and the waiting behavior on fuchsia happens in fdio, we can skip it for
ns3 for now.

Bug: 128641
Change-Id: I4f6893c1c96e7b70f79daacd3b1a0a4dcee85f8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872020
Reviewed-by: Nick Brown <nickbrow@google.com>
Commit-Queue: Zeling Feng <zeling@google.com>