[minimal] Add buildbot bundle for minimal
Add a bundle for minimal buildbots similar to other products.

Include an "additional_build_targets" group that can be used to build
other reachable GN targets that are not dependencies of the image and
thus are not built by `fx build` without the bundle. The initial set
of targets is a manual selection of targets that are known and build
successfully on `minimal.riscv64`.

Bug: 128868
Change-Id: Id466809d369993f16f799a676441451ef3044d76
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871619
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: John Wittrock <wittrock@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>